% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_colors.R
\name{team_colors}
\alias{team_colors}
\title{Generate NCAA football team color palettes}
\usage{
team_colors(team_name, n = length(palette))
}
\arguments{
\item{team_name}{Character string of school name}

\item{n}{Numeric indicating number of colors returned; defaults to all}
}
\description{
Return colors for a specific team.
}
