% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_colors.R
\docType{data}
\name{teams}
\alias{teams}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 130.
}
\usage{
teams
}
\description{
Complete list of palettes
}
\keyword{datasets}
